refactor: Changed files organization for services
