+++ src/app/components/dashboard/dashboard.component.ts
@@ -1,4 +1,4 @@
-import { AuthService } from "src/app/services/auth.service";
+import { AuthService } from "../../services/auth/auth.service";
 import { Component } from "@angular/core";
 
 @Component({
